Starting base for the Main Brand Panel with Fonts
Created initial Main Brand Panel with a fonts section and made this panel the index.tsx. Then, made the old index.tsx file the ColorPanel.tsx file.

Linked the "Manage brand palette" button to open ColorPanel.tsx
